+++ marstek_bridge/app/__init__.py
@@ -1,4 +1,4 @@
 """Marstek for Self-consumption regulation (MQTT - UDP Bridge)."""
 
-__version__ = "0.0.7"
+__version__ = "0.0.8"
 __project__ = "Marstek for Self-consumption regulation (MQTT - UDP Bridge)"